backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.2
python-multipart
httpx
pytest
pytest-asyncio
